backend/requirements.txt
beautifulsoup4==4.12.2
Django==5.0.1
djangorestframework==3.14.0
selenium==4.16.0
webdriver_manager==4.0.1
